Replace Flask with FastAPI
Since Flask showed poor results when benchmarked with wrk, this commit
replaces it with FastAPI, which led to much better results.
To avoid breaking the existing API of the visitors module, we are
re-serializing the log entry after the initial de-serialization
performed automatically by FastAPI. This should be changed in the future
if the performance impact becomes too high.